split up testing and functionality to focus on Downloader object
