+++ sec_web_scraper/__main__.py
@@ -1,4 +1,5 @@
-from Scraper import get_company_filings_given_cik, iterate_over_filings, build_index_sec
+from Scraper import get_company_filings_given_cik, iterate_over_filings
+from Downloader import Downloader
 
 
 def main():
@@ -20,7 +21,10 @@ def main():
     print(f"This is the file count {n}")
     print(sample_10k)
 
-    build_index_sec(2010, 2012)  # Small Sample of Documents
+    d = Downloader()
+    d.build_index_sec(2010, 2010)  # Small Sample of Documents
+    d.pretty_print_forms()
+    d.read_tsv_files()
 
 
 if __name__ == "__main__":
